disable indexing_slicing for custom Index impls
